refactor(#26): 뜻 가리는 버튼 컴포넌트로 분리 후 적용
