+++ features/hide-word/index.ts
@@ -1,0 +1,3 @@
+export { useHideWordsMeaning } from './model/useHideWordsMeaning'
+
+export { default as HideWordButton } from './ui/HideWordButton'
