+++ features/hide-word/model/useHideWordsMeaning.ts
@@ -1,0 +1,15 @@
+import { useState } from 'react'
+
+export function useHideWordsMeaning() {
+  const [isHidden, setIsHidden] = useState(false)
+
+  const handleHideMeaning = () => {
+    setIsHidden(!isHidden)
+  }
+
+  return {
+    isHidden,
+    setIsHidden,
+    handleHideMeaning,
+  }
+}
